+++ ras-x86-repr/src/operand.rs
@@ -3,15 +3,15 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Copy, Debug, PartialEq, Serialize, Deserialize)]
 pub struct OperandRepr {
     pub kind: OperandKind,
-    pub size: usize,
+    pub size: u32,
 }
 
 impl OperandRepr {
-    pub fn new(kind: OperandKind, size: usize) -> Self {
+    pub fn new(kind: OperandKind, size: u32) -> Self {
         Self { kind, size }
     }
 
-    pub fn size(&self) -> usize {
+    pub fn size(&self) -> u32 {
         self.size
     }
 }
